+++ envergo/moulinette/tests/test_models.py
@@ -98,154 +98,3 @@ def test_moulinette_get_perimeters_distance(
     perimeter.activation_distance = 60
     perimeter.save()
     assert len(moulinette.get_perimeters()) == 4
-
-
-@pytest.mark.parametrize("footprint", [50])
-def test_3310_small_footprint_outside_wetlands(moulinette_data):
-    """Project with footprint < 700m² are not subject to the 3310."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_soumis"
-
-
-@pytest.mark.parametrize("footprint", [50])
-def test_3310_small_footprint_inside_wetlands(moulinette_data):
-    """Project with footprint < 700m² are not subject to the 3310."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_soumis"
-
-
-@pytest.mark.parametrize("footprint", [700])
-def test_3310_medium_footprint_inside_wetlands(moulinette_data):
-    """Project with 700 <= footprint <= 1000m² within a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "action_requise"
-
-
-@pytest.mark.parametrize("footprint", [800])
-def test_3310_medium_footprint_inside_wetlands_2(moulinette_data):
-    """Project with 700 <= footprint <= 1000m² within a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "action_requise"
-
-
-@pytest.mark.parametrize("footprint", [800])
-def test_3310_medium_footprint_close_to_wetlands(moulinette_data):
-    """Project with 700 <= footprint <= 1000m² close to a wetland."""
-
-    # Make sure the project in close to a wetland
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = False
-    moulinette.catalog["wetlands_within_100m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_soumis"
-
-
-@pytest.mark.parametrize("footprint", [800])
-def test_3310_medium_footprint_inside_potential_wetlands(moulinette_data):
-    """Project with 700 <= footprint <= 1000m² inside a potential wetland."""
-
-    # Make sure the project is in a potential wetland
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = False
-    moulinette.catalog["wetlands_within_100m"] = False
-    moulinette.catalog["within_potential_wetlands"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_soumis"
-
-
-@pytest.mark.parametrize("footprint", [800])
-def test_3310_medium_footprint_outside_wetlands(moulinette_data):
-    """Project with 700 < footprint < 1000m² outside a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_concerne"
-
-
-@pytest.mark.parametrize("footprint", [1500])
-def test_3310_large_footprint_within_wetlands(moulinette_data):
-    """Project with footprint >= 1000m² within a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "soumis"
-
-
-@pytest.mark.parametrize("footprint", [1500])
-def test_3310_large_footprint_close_to_wetlands(moulinette_data):
-    """Project with footprint >= 1000m² close to a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = False
-    moulinette.catalog["wetlands_within_100m"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "action_requise"
-
-
-@pytest.mark.parametrize("footprint", [1500])
-def test_3310_large_footprint_inside_potential_wetland(moulinette_data):
-    """Project with footprint >= 1000m² inside a potential wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["wetlands_within_25m"] = False
-    moulinette.catalog["wetlands_within_100m"] = False
-    moulinette.catalog["within_potential_wetlands"] = True
-    assert moulinette.loi_sur_leau.zone_humide.result == "action_requise"
-
-
-@pytest.mark.parametrize("footprint", [1500])
-def test_3310_large_footprint_outside_wetlands(moulinette_data):
-    """Project with footprint > 1000m² outside a wetland."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    assert moulinette.loi_sur_leau.zone_humide.result == "non_concerne"
-
-
-@pytest.mark.parametrize("footprint", [299])
-def test_3220_small_footprint(moulinette_data):
-    """Project with footprint < 300m² are not subject to the 3320."""
-
-    # Make sure the project in in a flood zone
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["flood_zones_within_12m"] = True
-    assert moulinette.loi_sur_leau.zone_inondable.result == "non_soumis"
-
-
-@pytest.mark.parametrize("footprint", [300])
-def test_3220_medium_footprint_within_flood_zones(moulinette_data):
-    """Project with 500m² < footprint <= 300m² within a flood zone."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["flood_zones_within_12m"] = True
-    assert moulinette.loi_sur_leau.zone_inondable.result == "action_requise"
-
-
-@pytest.mark.parametrize("footprint", [300])
-def test_3220_medium_footprint_outside_flood_zones(moulinette_data):
-    """Project with 500m² < footprint <= 300m² outside a flood zone."""
-
-    # Make sure the project in in a flood zone
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    assert moulinette.loi_sur_leau.zone_inondable.result == "non_concerne"
-
-
-@pytest.mark.parametrize("footprint", [400])
-def test_3220_large_footprint_within_flood_zones(moulinette_data):
-    """Project with footprint >= 400m² within a flood zone."""
-
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    moulinette.catalog["flood_zones_within_12m"] = True
-    assert moulinette.loi_sur_leau.zone_inondable.result == "soumis"
-
-
-@pytest.mark.parametrize("footprint", [300])
-def test_3220_large_footprint_outside_flood_zones(moulinette_data):
-    """Project with footprint >= 400m² outside a flood zone."""
-
-    # Make sure the project in in a flood zone
-    moulinette = Moulinette(moulinette_data, moulinette_data)
-    assert moulinette.loi_sur_leau.zone_inondable.result == "non_concerne"
